Move `OpenHours`, `OpenPeriod`, `WeekDay`, `WeekTimestamp`, `ParkingData` to common module
