+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/history/HistoryRecordsSerializerTest.kt
@@ -7,11 +7,11 @@ import com.mapbox.search.SearchResultMetadata
 import com.mapbox.search.TestData
 import com.mapbox.search.base.result.BaseRawResultType
 import com.mapbox.search.common.RoutablePoint
-import com.mapbox.search.metadata.OpenHours
-import com.mapbox.search.metadata.OpenPeriod
-import com.mapbox.search.metadata.ParkingData
-import com.mapbox.search.metadata.WeekDay
-import com.mapbox.search.metadata.WeekTimestamp
+import com.mapbox.search.common.metadata.OpenHours
+import com.mapbox.search.common.metadata.OpenPeriod
+import com.mapbox.search.common.metadata.ParkingData
+import com.mapbox.search.common.metadata.WeekDay
+import com.mapbox.search.common.metadata.WeekTimestamp
 import com.mapbox.search.record.HistoryRecord
 import com.mapbox.search.record.HistoryRecordsSerializer
 import com.mapbox.search.result.SearchAddress
