+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/utils/serialization/WeekDayDAO.kt
@@ -1,7 +1,7 @@
 package com.mapbox.search.utils.serialization
 
 import com.google.gson.annotations.SerializedName
-import com.mapbox.search.metadata.WeekDay
+import com.mapbox.search.common.metadata.WeekDay
 
 internal enum class WeekDayDAO : DataAccessObject<WeekDay> {
